Extracted main logic to Domain Object
