+++ app/controllers/role_lists_controller.rb
@@ -12,23 +12,23 @@ class RoleListsController < CrudController
 
   helper_method :group
 
-  respond_to :js, only: [:move, :movable, :deletable]
+  respond_to :js, only: [:new, :move, :movable, :deletable]
 
   def create
-    new_roles = build_new_roles_hash
+    new_roles = role_list.build_new_roles_hash
     count = Role.create(new_roles).count
     redirect_to(group_people_path(group), notice: flash_message(:success, count: count))
   end
 
   def destroy
-    count = Role.destroy(deletable_role_ids).count
+    count = Role.destroy(role_list.deletable_role_ids).count
     redirect_to(group_people_path(group), notice: flash_message(:success, count: count))
   end
 
   def update
     count = Role.transaction do
-      Role.destroy(deletable_role_ids)
-      Role.create(build_moving_roles_hash).count
+      Role.destroy(role_list.deletable_role_ids)
+      Role.create(role_list.build_new_roles_hash).count
     end
 
     redirect_to(group_people_path(group), notice: flash_message(:success, count: count))
@@ -47,18 +47,14 @@ class RoleListsController < CrudController
 
   def movable
     @new_group_id = model_params[:group_id]
-    @new_role_type = model_params[:type]
-    @role_types = collect_available_role_types
+    @new_role_type = role_type
+    @role_types = role_list.collect_available_role_types
     @people_ids = params[:ids]
   end
 
   def deletable
-    @role_types = collect_available_role_types
+    @role_types = role_list.collect_available_role_types
     @people_ids = params[:ids]
-  end
-
-  def entry
-    super.decorate
   end
 
   def self.model_class
@@ -67,16 +63,8 @@ class RoleListsController < CrudController
 
   private
 
-  def collect_available_role_types
-    roles.each_with_object({}) do |role, hash|
-      next unless can?(:destroy, role)
-      key = role.group.name
-      hash[key] = {} if hash[key].blank?
-
-      type = role.type
-      count = hash[key][type].blank? ? 1 : hash[key][type] + 1
-      hash[key][type] = count
-    end
+  def entry
+    super.decorate
   end
 
   def validate_role_type
@@ -85,50 +73,8 @@ class RoleListsController < CrudController
     end
   end
 
-  def build_new_roles_hash
-    people.map do |person|
-      role = build_role(person.id)
-      authorize!(:create, role, message: access_denied_flash(person))
-      role.attributes
-    end
-  end
-
-  def build_moving_roles_hash
-    people.map do |person|
-      new_role = build_role(person.id)
-      authorize!(:create, new_role, message: access_denied_flash(person))
-      new_role.attributes
-    end
-  end
-
-  def build_role(person_id)
-    role = build_role_type
-    role.attributes = permitted_params(@type)
-    role.person_id = person_id
-    role
-  end
-
-  def build_role_type
-    @type = new_group.class.find_role_type!(role_type)
-    @type.new
-  end
-
-  def deletable_role_ids
-    roles.where(type: role_types.keys).map do |r|
-      authorize!(:destroy, r, message: access_denied_flash(r.person)).id
-    end
-  end
-
-  def permitted_params(role_type = entry.class)
-    model_params.permit(role_type.used_attributes + permitted_attrs)
-  end
-
   def handle_access_denied(e)
     redirect_to(group_people_path(group), alert: e.message)
-  end
-
-  def access_denied_flash(person)
-    I18n.t("#{controller_name}.access_denied", person: person.full_name)
   end
 
   def flash_message(type, attrs = {})
@@ -137,14 +83,6 @@ class RoleListsController < CrudController
 
   def role_type
     model_params[:type]
-  end
-
-  def role_types
-    model_params && model_params[:types] ? model_params[:types] : {}
-  end
-
-  def new_group
-    @new_group ||= Group.find(model_params[:group_id])
   end
 
   def group
@@ -159,16 +97,11 @@ class RoleListsController < CrudController
     params[:ids].to_s.split(',')
   end
 
-  def roles
-    group_ids = params[:range] == ('layer' || 'deep') ? layer_group_ids : group
-    @roles ||= Role.where(person_id: people_ids, group_id: group_ids)
-  end
-
-  def layer_group_ids
-    group.groups_in_same_layer.pluck(:id)
-  end
-
   def person_filter
     @person_filter ||= Person::Filter::List.new(@group, current_user, list_filter_args)
+  end
+
+  def role_list
+    @role_list ||= Role::List.new(current_ability, params)
   end
 end
